option --sourcetgz: use git repo from local file
bootstrap-perl --sourcetgz /var/tmp/perl.tgz

will use this file as initial git repo.
This speeds up things a lot and helps
when running continuous Perl builds.

After unpacking this file it will do
a standard git pull to update the repo
so you don't need to update the .tgz.

It MUST be a tar gz file for simplicity yet.