Rework injection of metainfo into %Config
Instead of directly appending to the generated
config.sh we now generate a config.arch which
in turn appends to UU/config.sh during Configure's
run, so that later gets merged into the ultimate config.

This is replacing one dirty mechanism with another
dirty mechanism, however, the merging of values
is then done by Configure, so it is a more clean
way, reducing the risk that bootstrap meta information
interferes with the real Perl config values.

Thanks to Tux for help sorting this out.